[milling] passaggio completo al processing ricorsivo dell'albero ovvero implementata Octree::processTree TODO: multithreaded processing

+++ source/common/EulerAngles.cpp
@@ -11,7 +11,7 @@ EulerAngles::EulerAngles() :
 	ALPHA(0), BETA(0), GAMMA(0) {
 }
 
-EulerAngles::EulerAngles(float alpha, float beta, float gamma) :
+EulerAngles::EulerAngles(double alpha, double beta, double gamma) :
 	ALPHA(alpha), BETA(beta), GAMMA(gamma) {
 }
